Duplicated mapfull.ajaxUrl as appsetup.env.api.url
